Fix failure to encode a list-like table with string keys
Problem:
A table can't be encoded, if its keys converted to numbers are unique
integers in interval [1; n], and some of them are strings. A table like
that is treated like an array. `lmpack_objlen` uses `lua_isnumber` to
determine if a key is a number. `lua_isnumber` returns
1 for a string convertible to a number.

Solution:
Use `lua_type` instead.